requirements.txt
folium==0.18.0
geopy==2.4.1
Pillow==11.2.1
streamlit==1.40.2
streamlit_folium==0.23.2
